refactor: complete Phase 4.3 Move Method - improve class cohesion and responsibility distribution
Phase 4.3 - Move Method refactoring:
- Move setupClientCompatibility() from server.ts to McpClientCompatibility
  * Improves client compatibility logic cohesion
  * Removes feature envy from server.ts
  * Centralizes client setup operations in appropriate class
- Move isUnityAvailable() from unity-discovery.ts to UnityClient (static method)
  * Consolidates low-level TCP connection testing in UnityClient
  * Eliminates duplicate connection test functionality
  * Removes unused net import from unity-discovery.ts
- Move fetchToolDetailsFromUnity() from unity-tool-manager.ts to UnityClient
  * Unifies Unity communication logic in UnityClient
  * Allows UnityToolManager to focus on tool management responsibilities
  * Improves parameter handling with explicit includeDevelopmentOnly flag

Improved class cohesion, eliminated feature envy patterns, and enhanced separation of concerns following Martin Fowler's Move Method refactoring principles.
